requirements.txt
black==22.3.0
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.2
